Added survey page
Added a page to the components folder that contains a survey to be completed by the user upon finishing a recipe. The page is currently incomplete, as I will be working with the rest of the team to fix some issues and figure out how to make the necessary components.