Pages kinda sorta work!
Fixed the error that prevented the user from navigating to the progression page. Besides figuring out how to use useContext, I also discovered the main issue was with the CheckBox object that I had imported. I will look into a fix for that at a later time, so for now my next objective is to make the page look good.