Added Functionality to RecipeSurvey
The survey page is (mostly) done! Every step, as well as the ingredient gathering process, has a trio of buttons associated with it. Each trio is independent from one another, and only one option can be selected for each. Once every trio has an option selected, the submit button becomes usable and it takes you to the skills page.

Currently, however, this page does NOT update the user's skills. This will be done in the next sprint most likely. In addition, the page still does not look very good due to me focusing on functionality over style.