Styled Recipe Page
Added styling to the recipe description and progression pages, as well as connected them to the global stylesheet. Only issue with it is that the "next" button in the step list is not going to the bottom of the page. This section also needs a back button like the ingredient list.