Fixed Timer
Changed the timer on recipe progression to use a simple delay function instead of setInterval, greatly reducing the lag it caused. It has two minor issues (it takes one second to start and only pauses after the next delay happens), but they are negligible problems.

Also added a blurb to the favorites page for when a user has nothing in their favorites list.