Work on Completed.js and Keywords
Added the API calls needed for completed.js to display the user's completed recipes. This does not work for some reason, will debug tomorrow.

Also added the remaining setup code for the keywords system. Need to add an API call that gets the keyword suffix data for the page, and need to connect the resulting Pressables to the Modal.