Updated timer
Timer now has a reset button, and is turned off by default. It also now displays "Start" when the timer is set to 0 seconds. Reset button is also disabled if the timer is running to prevent misuse.

It still has a weird issue where the timer lags by a second when starting and when pausing.